renamed testing environment script to exclude it from runned tests
