+++ test/app.test.js
@@ -1,4 +1,4 @@
-const { test, superTest } = require('./test-env');
+const { test, superTest } = require('./test-env/_test_env');
 
 test('test root route', async (t) => {
   const { statusCode } = await superTest.get('/');
