+++ test/contract/cloudinary-upload.test.js
@@ -1,11 +1,9 @@
-const { test, superTest } = require('../test-env');
+const { test, superTest } = require('../test-env/_test_env');
 const path = require('path');
 
 test('test upload file cloudinary', async (t) => {
   const resp = await superTest.post('/cloudinary/upload')
     .attach('input', path.join(__dirname, '..', '/tmp/1.jpg'));
-
-    console.log(resp);
 
   t.is(resp.statusCode, 201);
 });
