+++ test/test-env/_test_env.js
@@ -5,8 +5,4 @@ const superTest = require('supertest-promised')(app);
 const test = require('ava');
 
 
-test('dummy test', (t) => {
-  t.is(true, true);
-});
-
 module.exports = { superTest, app, test };
